[refactor/#58] 클래스 분리
