+++ src/main/java/at/mateball/domain/groupmember/core/repository/GroupMemberRepository.java
@@ -1,0 +1,10 @@
+package at.mateball.domain.groupmember.core.repository;
+
+import at.mateball.domain.groupmember.core.GroupMember;
+import at.mateball.domain.groupmember.core.repository.querydsl.GroupMemberRepositoryCustom;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface GroupMemberRepository extends JpaRepository<GroupMember, Long>, GroupMemberRepositoryCustom {
+}
