+++ src/main/java/at/mateball/domain/groupmember/core/repository/querydsl/GroupMemberRepositoryCustom.java
@@ -1,0 +1,11 @@
+package at.mateball.domain.groupmember.core.repository.querydsl;
+
+
+import java.util.List;
+import java.util.Map;
+
+public interface GroupMemberRepositoryCustom {
+    Map<Long, Integer> findGroupMemberCountMap(List<Long> groupIds);
+
+    Map<Long, List<String>> findGroupMemberImgMap(List<Long> groupIds);
+}
